+++ src/aux-common/bots/BotCalculations.ts
@@ -2396,10 +2396,31 @@ export function convertToCopiableValue(value: any): any {
 }
 
 export function getCreatorVariable(context: BotSandboxContext, bot: ScriptBot) {
+    return getBotVariable(context, bot, 'auxCreator');
+}
+
+export function getConfigVariable(context: BotSandboxContext, bot: ScriptBot) {
+    return getBotVariable(context, bot, 'auxConfigBot');
+}
+
+export function getConfigTagVariable(
+    context: BotSandboxContext,
+    bot: ScriptBot,
+    tag: keyof BotTags,
+    config: ScriptBot
+) {
+    return config && tag ? config.tags[tag] : null;
+}
+
+export function getBotVariable(
+    context: BotSandboxContext,
+    bot: ScriptBot,
+    tag: string
+): ScriptBot {
     if (!bot) {
         return null;
     }
-    let creatorId = context.sandbox.interface.getTag(bot, 'auxCreator');
+    let creatorId = context.sandbox.interface.getTag(bot, tag);
     if (creatorId) {
         let obj = context.sandbox.interface.getBot(creatorId);
         if (obj) {
@@ -2430,11 +2451,16 @@ function _calculateFormulaValue(
     const scriptBot = getScriptBot(context, object);
     setCurrentBot(scriptBot);
 
+    const creator = getCreatorVariable(context, scriptBot);
+    const config = getConfigVariable(context, scriptBot);
     let vars = {
         bot: scriptBot,
         tags: scriptBot ? scriptBot.tags : null,
         raw: scriptBot ? scriptBot.raw : null,
-        creator: getCreatorVariable(context, scriptBot),
+        tagName: tag || null,
+        creator: creator,
+        config: config,
+        configTag: getConfigTagVariable(context, scriptBot, tag, config),
     };
 
     // NOTE: The energy should not get reset